// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Plans = {
  STANDARD: 'STANDARD',
  PRO: 'PRO',
  ULTIMATE: 'ULTIMATE'
} as const

export type Plans = (typeof Plans)[keyof typeof Plans]


export const Role = {
  OWNER: 'OWNER',
  CUSTOMER: 'CUSTOMER',
  assistant: 'assistant',
  user: 'user'
} as const

export type Role = (typeof Role)[keyof typeof Role]